requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
certifi==2023.7.22
charset-normalizer==3.3.2
distro==1.9.0
docker-pycreds==0.4.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
multidict==6.0.5
munkres==1.1.4
openai==0.28.0
pathtools==0.1.2
protobuf==4.23.4
pydantic==2.6.3
pydantic_core==2.16.3
python-dotenv==1.0.1
requests==2.31.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
ruamel.yaml.string==0.1.1
sniffio==1.3.1
tenacity==8.2.3
tqdm==4.66.2
urllib3==2.2.1
yarl==1.9.4
